+++ packages/next/src/utils.ts
@@ -192,8 +192,12 @@ function normalizePage(page: string): string {
   if (!page.startsWith('/')) {
     page = `/${page}`;
   }
-  // remove '/index' from the end
-  page = page.replace(/\/index$/, '/');
+
+  // Replace the `/index` with `/`
+  if (page === '/index') {
+    page = '/';
+  }
+
   return page;
 }
 
@@ -320,8 +324,8 @@ export async function getDynamicRoutes({
   bypassToken,
   isServerMode,
   dynamicMiddlewareRouteMap,
-  experimentalPPRRoutes,
   hasActionOutputSupport,
+  isAppPPREnabled,
 }: {
   entryPath: string;
   entryDirectory: string;
@@ -333,8 +337,8 @@ export async function getDynamicRoutes({
   bypassToken?: string;
   isServerMode?: boolean;
   dynamicMiddlewareRouteMap?: ReadonlyMap<string, RouteWithSrc>;
-  experimentalPPRRoutes: ReadonlySet<string>;
   hasActionOutputSupport: boolean;
+  isAppPPREnabled: boolean;
 }): Promise<RouteWithSrc[]> {
   if (routesManifest) {
     switch (routesManifest.version) {
@@ -408,7 +412,7 @@ export async function getDynamicRoutes({
             ];
           }
 
-          if (experimentalPPRRoutes.has(page)) {
+          if (isAppPPREnabled) {
             let dest = route.dest?.replace(/($|\?)/, '.prefetch.rsc$1');
 
             if (page === '/' || page === '/index') {
@@ -1439,6 +1443,13 @@ async function getSourceFilePathFromPage({
     }
   }
 
+  // if we got here, and didn't find a source not-found file, then it was the one injected
+  // by Next.js. There's no need to warn or return a source file in this case, as it won't have
+  // any configuration applied to it.
+  if (page === '/_not-found/page') {
+    return '';
+  }
+
   console.log(
     `WARNING: Unable to find source file for page ${page} with extensions: ${extensionsToTry.join(
       ', '
@@ -1497,9 +1508,9 @@ export type LambdaGroup = {
   maxDuration?: number;
   isAppRouter?: boolean;
   isAppRouteHandler?: boolean;
-  isStreaming?: boolean;
-  isPrerenders?: boolean;
-  isExperimentalPPR?: boolean;
+  readonly isStreaming: boolean;
+  readonly isPrerenders: boolean;
+  readonly isExperimentalPPR: boolean;
   isActionLambda?: boolean;
   isPages?: boolean;
   isApiLambda: boolean;
@@ -1554,6 +1565,7 @@ export async function getPageLambdaGroups({
     const routeName = normalizePage(page.replace(/\.js$/, ''));
     const isPrerenderRoute = prerenderRoutes.has(routeName);
     const isExperimentalPPR = experimentalPPRRoutes?.has(routeName) ?? false;
+    const isStreaming = !isPrerenderRoute || isExperimentalPPR;
 
     let opts: { memory?: number; maxDuration?: number } = {};
 
@@ -1565,15 +1577,13 @@ export async function getPageLambdaGroups({
     }
 
     if (config && config.functions) {
-      // `pages` are normalized without route groups (e.g., /app/(group)/page.js).
-      // we keep track of that mapping in `inversedAppPathManifest`
-      // `getSourceFilePathFromPage` needs to use the path from source to properly match the config
-      const pageFromManifest = inversedAppPathManifest?.[routeName];
       const sourceFile = await getSourceFilePathFromPage({
         workPath: entryPath,
-        // since this function is used by both `pages` and `app`, the manifest might not be provided
-        // so fallback to normal behavior of just checking the `page`.
-        page: pageFromManifest ?? page,
+        page: normalizeSourceFilePageFromManifest(
+          routeName,
+          page,
+          inversedAppPathManifest
+        ),
         pageExtensions,
       });
 
@@ -1627,6 +1637,7 @@ export async function getPageLambdaGroups({
         ...opts,
         isPrerenders: isPrerenderRoute,
         isExperimentalPPR,
+        isStreaming,
         isApiLambda: !!isApiPage(page),
         pseudoLayerBytes: initialPseudoLayer.pseudoLayerBytes,
         pseudoLayerUncompressedBytes: initialPseudoLayerUncompressed,
@@ -1655,6 +1666,43 @@ export async function getPageLambdaGroups({
   }
 
   return groups;
+}
+
+// `pages` are normalized without route groups (e.g., /app/(group)/page.js).
+// we keep track of that mapping in `inversedAppPathManifest`
+// `getSourceFilePathFromPage` needs to use the path from source to properly match the config
+function normalizeSourceFilePageFromManifest(
+  routeName: string,
+  page: string,
+  inversedAppPathManifest?: Record<string, string>
+) {
+  const pageFromManifest = inversedAppPathManifest?.[routeName];
+  if (!pageFromManifest) {
+    // since this function is used by both `pages` and `app`, the manifest might not be provided
+    // so fallback to normal behavior of just checking the `page`.
+    return page;
+  }
+
+  const metadataConventions = [
+    '/favicon.',
+    '/icon.',
+    '/apple-icon.',
+    '/opengraph-image.',
+    '/twitter-image.',
+    '/sitemap.',
+    '/robots.',
+  ];
+
+  // these special metadata files for will not contain `/route` or `/page` suffix, so return the routeName as-is.
+  const isSpecialFile = metadataConventions.some(convention =>
+    routeName.startsWith(convention)
+  );
+
+  if (isSpecialFile) {
+    return routeName;
+  }
+
+  return pageFromManifest;
 }
 
 export const outputFunctionFileSizeInfo = (
@@ -2177,22 +2225,6 @@ export const onPrerenderRoute =
       initialStatus = 404;
     }
 
-    /**
-     * If the route key had an `/index` suffix added, we need to note it so we
-     * can remove it from the output path later accurately.
-     */
-    let addedIndexSuffix = false;
-
-    if (isAppPathRoute) {
-      // for literal index routes we need to append an additional /index
-      // due to the proxy's normalizing for /index routes
-      if (routeKey !== '/index' && routeKey.endsWith('/index')) {
-        routeKey = `${routeKey}/index`;
-        routeFileNoExt = routeKey;
-        addedIndexSuffix = true;
-      }
-    }
-
     let outputPathPage = path.posix.join(entryDirectory, routeFileNoExt);
 
     if (!isAppPathRoute) {
@@ -2386,7 +2418,7 @@ export const onPrerenderRoute =
         // static route first, then try the srcRoute if it doesn't exist. If we
         // can't find it at all, this constitutes an error.
         experimentalStreamingLambdaPath = experimentalStreamingLambdaPaths.get(
-          pathnameToOutputName(entryDirectory, routeKey, addedIndexSuffix)
+          pathnameToOutputName(entryDirectory, routeKey)
         );
         if (!experimentalStreamingLambdaPath && srcRoute) {
           experimentalStreamingLambdaPath =
@@ -2465,6 +2497,7 @@ export const onPrerenderRoute =
           ...(rscEnabled
             ? {
                 initialHeaders: {
+                  ...initialHeaders,
                   'content-type': rscContentTypeHeader,
                   vary: rscVaryHeader,
                   // If it contains a pre-render, then it was postponed.
@@ -2475,6 +2508,22 @@ export const onPrerenderRoute =
               }
             : {}),
         });
+      }
+
+      // we need to ensure all prerenders have a matching .rsc output
+      // otherwise routing could fall through unexpectedly for the
+      // fallback: false case as it doesn't have a dynamic route
+      // to catch the `.rsc` request for app -> pages routing
+      if (outputPrerenderPathData?.endsWith('.json') && appDir) {
+        const dummyOutput = new FileBlob({
+          data: '{}',
+          contentType: 'application/json',
+        });
+        const rscKey = `${outputPathPage}.rsc`;
+        const prefetchRscKey = `${outputPathPage}${RSC_PREFETCH_SUFFIX}`;
+
+        prerenders[rscKey] = dummyOutput;
+        prerenders[prefetchRscKey] = dummyOutput;
       }
 
       ++prerenderGroup;
@@ -2630,18 +2679,9 @@ export function getNextServerPath(nextVersion: string) {
     : 'next/dist/next-server/server';
 }
 
-function pathnameToOutputName(
-  entryDirectory: string,
-  pathname: string,
-  addedIndexSuffix = false
-) {
+function pathnameToOutputName(entryDirectory: string, pathname: string) {
   if (pathname === '/') {
     pathname = '/index';
-  }
-  // If the `/index` was added for a route that ended in `/index` we need to
-  // strip the second one off before joining it with the entryDirectory.
-  else if (addedIndexSuffix) {
-    pathname = pathname.replace(/\/index$/, '');
   }
 
   return path.posix.join(entryDirectory, pathname);
@@ -2768,7 +2808,7 @@ interface EdgeFunctionInfoV2 extends BaseEdgeFunctionInfo {
 
 interface EdgeFunctionInfoV3 extends BaseEdgeFunctionInfo {
   matchers: EdgeFunctionMatcher[];
-  environments: Record<string, string>;
+  env: Record<string, string>;
 }
 
 interface EdgeFunctionMatcher {
@@ -3003,7 +3043,7 @@ export async function getMiddlewareBundle({
                   slug: 'nextjs',
                   version: nextVersion,
                 },
-                environment: edgeFunction.environments,
+                environment: edgeFunction.env,
               });
             })(),
             routeMatchers: getRouteMatchers(edgeFunction, routesManifest),
@@ -3183,7 +3223,7 @@ export function upgradeMiddlewareManifestV1(
     return {
       ...rest,
       matchers: [{ regexp }],
-      environments: {},
+      env: {},
     };
   }
 
@@ -3211,7 +3251,7 @@ export function upgradeMiddlewareManifestV2(
     const { ...rest } = v2Info;
     return {
       ...rest,
-      environments: {},
+      env: {},
     };
   }